Potentially fix the rest of the derive macros for generic structs
(Drawing the rest of the owl)

The workaround for DebugBits and DefaultBits is a little overkill,
but much more trivial than trying to figure out whether we need to
emit the where clause for each field. We should be able to filter
it in the future if necessary.

Perhaps this should also be refactored into a shared helper function,
since it was copy-pasted into two locations.